# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mad Factory - Factory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/statement:
    get:
      operationId: getCustomerStatement
      tags: [customers]
      summary: Get customer account statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customer account statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStatement"
  /customers/{id}/opening-balance:
    post:
      operationId: setCustomerOpeningBalance
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpeningBalanceBody"
      responses:
        "200":
          description: Updated

  # ── Suppliers ──────────────────────────────────────────────────────────────
...
  /customer-payments:
    get:
      operationId: listCustomerPayments
      tags: [payments]
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customer payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerPayment"
    post:
      operationId: createCustomerPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerPaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPayment"
  /customer-payments/{id}:
    get:
      operationId: getCustomerPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPayment"
    delete:
      operationId: deleteCustomerPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Purchase Invoices (IQD) ────────────────────────────────────────────────
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # ── Payroll ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        phone2:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        openingBalance:
          type: number
        totalDebt:
          type: number
        totalPaid:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - isActive
        - openingBalance
        - totalDebt
        - totalPaid
        - createdAt
        - updatedAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        phone2:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        openingBalance:
          type: number
      required:
        - name
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        phone2:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Supplier ───────────────────────────────────────────────────────────────
...
    CustomerPayment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerId
        - customerName
        - amount
        - paymentDate
        - createdAt
    CreateCustomerPaymentBody:
      type: object
      properties:
        customerId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        voucherType:
          type: string
          enum: [cash, internal]
        notes:
          type: ["string", "null"]
      required:
        - customerId
        - amount
        - paymentDate

    # ── Purchase Invoice ──────────────────────────────────────────────────────
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        role:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        phone2:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        guarantor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        salary:
          type: number
        isActive:
          type: boolean
        hireDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - code
        - name
        - salary
        - isActive
        - createdAt
        - updatedAt
    CreateEmployeeBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        role:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        phone2:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        guarantor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        salary:
          type: number
        hireDate:
          type: ["string", "null"]
      required:
        - code
        - name
        - salary
    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        phone2:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        guarantor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        salary:
          type: number
        isActive:
          type: boolean
        hireDate:
          type: ["string", "null"]

    # ── Payroll ───────────────────────────────────────────────────────────────
...
    StatementEntry:
      type: object
      properties:
        date:
          type: string
        type:
          type: string
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
        balance:
          type: number
        referenceId:
          type: ["integer", "null"]
      required:
        - date
        - type
        - description
        - debit
        - credit
        - balance
    AccountStatement:
      type: object
      properties:
        entityId:
          type: integer
        entityName:
          type: string
        openingBalance:
          type: number
        totalDebit:
          type: number
        totalCredit:
          type: number
        closingBalance:
          type: number
        entries:
          type: array
          items:
            $ref: "#/components/schemas/StatementEntry"
      required:
        - entityId
        - entityName
        - openingBalance
        - totalDebit
        - totalCredit
        - closingBalance
        - entries

    # ── Daily Report ──────────────────────────────────────────────────────────
...
    OpeningBalanceBody:
      type: object
      properties:
        amount:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - amount

    # ── App Settings ──────────────────────────────────────────────────────────
...